Moved the builder services to a subfolder.
