+++ source/components/cardContainer/builder/index.ts
@@ -1,0 +1,9 @@
+import { CardContainerBuilder } from './cardContainerBuilder.service';
+import { DataSourceBuilder } from './dataSourceBuilder.service';
+import { FilterBuilder } from './filterBuilder.service';
+
+export const BUILDER_PROVIDERS: any[] = [CardContainerBuilder, DataSourceBuilder, FilterBuilder];
+
+export * from './cardContainerBuilder.service';
+export * from './dataSourceBuilder.service';
+export * from './filterBuilder.service';
