+++ source/components/cardContainer/index.ts
@@ -1,3 +1,4 @@
+import * as builder from './builder/index';
 import * as container from './container/index';
 import * as dataSources from './dataSources/index';
 import * as filters from './filters/index';
@@ -6,10 +7,11 @@ import * as sorts from './sorts/index';
 import * as templates from './templates/index';
 
 export const CARD_CONTAINER_DIRECTIVES: any[] = [container.CONTAINER_DIRECTIVES, filters.FILTER_DIRECTIVES, templates.TEMPLATE_DIRECTIVES];
+export const CARD_CONTAINER_PROVIDERS: any[] = [builder.BUILDER_PROVIDERS];
 
-export { container, dataSources, filters, paging, sorts, templates };
+export { container, builder, dataSources, filters, paging, sorts, templates };
 
-// card
+export * from './card/card';
 export * from './cardContainer';
 export * from './cardContainerBuilder.service';
 export * from './column';
